trial 65: apply same-file memo-status detection to jsx-no-new-{array,object,jsx}-as-prop — extract shared util buildSameFileMemoRegistry / memoStatusForJsxOpeningName; same rationale as the function rule (React.memo bails only on memoised consumers)
