+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/react-builtins/jsx-no-new-function-as-prop.ts
@@ -4,6 +4,11 @@ import type { EsTreeNodeOfType } from "../../utils/es-tree-node-of-type.js";
 import { findVariableInitializer } from "../../utils/find-variable-initializer.js";
 import { isInsideFunctionScope } from "../../utils/is-inside-function-scope.js";
 import { isJsxAttributeOnIntrinsicHtmlElement } from "../../utils/is-on-intrinsic-html-element.js";
+import {
+  buildSameFileMemoRegistry,
+  memoStatusForJsxOpeningName,
+  type MemoStatus,
+} from "../../utils/build-same-file-memo-registry.js";
 import { isNodeOfType } from "../../utils/is-node-of-type.js";
 import { isTestlikeFilename } from "../../utils/is-testlike-filename.js";
 import { stripParenExpression } from "../../utils/strip-paren-expression.js";
@@ -748,85 +753,6 @@ const isParameterBindingWrapper = (expression: EsTreeNode): boolean => {
   return isStableStatementBlock(body.body ?? []);
 };
 
-// APPROACH 2: detect components that are KNOWN to NOT be memoised in
-// the current file. The rule's whole reason to fire is: "downstream
-// `React.memo` consumer breaks on the new reference". If we can prove
-// the consumer ISN'T `React.memo` / `forwardRef`-wrapped (because we
-// see its definition in the same file), the rule has nothing useful
-// to say.
-//
-// We don't do cross-file analysis — too expensive, too brittle —
-// but for SAME-FILE component definitions we can read the AST. The
-// registry maps component-name → "definitely memoised" / "definitely
-// NOT memoised" / "unknown (assume memoised, fire the rule)".
-type MemoStatus = "memoised" | "not-memoised" | "unknown";
-
-const HOC_NAMES_FOR_MEMOISATION: ReadonlySet<string> = new Set([
-  "memo",
-  "React.memo",
-  "forwardRef",
-  "React.forwardRef",
-  "observer", // MobX
-  "observable", // legend-state
-]);
-
-const flattenCalleeNameSimple = (callee: EsTreeNode): string | null => {
-  if (isNodeOfType(callee, "Identifier")) return callee.name;
-  if (
-    isNodeOfType(callee, "MemberExpression") &&
-    isNodeOfType(callee.object, "Identifier") &&
-    isNodeOfType(callee.property, "Identifier") &&
-    !callee.computed
-  ) {
-    return `${callee.object.name}.${callee.property.name}`;
-  }
-  return null;
-};
-
-const isMemoisingCall = (call: EsTreeNode): boolean => {
-  if (!isNodeOfType(call, "CallExpression")) return false;
-  const name = flattenCalleeNameSimple(call.callee as EsTreeNode);
-  return name !== null && HOC_NAMES_FOR_MEMOISATION.has(name);
-};
-
-const buildSameFileMemoRegistry = (program: EsTreeNode): Map<string, MemoStatus> => {
-  const registry = new Map<string, MemoStatus>();
-  if (!isNodeOfType(program, "Program")) return registry;
-  for (const statement of program.body) {
-    const root = isNodeOfType(statement as EsTreeNodeOfType<"ExportNamedDeclaration">, "ExportNamedDeclaration")
-      ? ((statement as EsTreeNodeOfType<"ExportNamedDeclaration">).declaration as EsTreeNode | null)
-      : isNodeOfType(statement as EsTreeNodeOfType<"ExportDefaultDeclaration">, "ExportDefaultDeclaration")
-        ? ((statement as EsTreeNodeOfType<"ExportDefaultDeclaration">).declaration as EsTreeNode | null)
-        : (statement as EsTreeNode);
-    if (!root) continue;
-    // `const X = memo(...)` / `const X = forwardRef(...)`
-    if (isNodeOfType(root, "VariableDeclaration")) {
-      for (const declarator of root.declarations ?? []) {
-        if (!isNodeOfType(declarator, "VariableDeclarator")) continue;
-        if (!isNodeOfType(declarator.id, "Identifier")) continue;
-        if (!declarator.init) continue;
-        const init = declarator.init as EsTreeNode;
-        if (isMemoisingCall(init)) {
-          registry.set(declarator.id.name, "memoised");
-        } else if (
-          isNodeOfType(init, "FunctionExpression") ||
-          isNodeOfType(init, "ArrowFunctionExpression")
-        ) {
-          // Plain function = definitely NOT memoised in this file.
-          registry.set(declarator.id.name, "not-memoised");
-        }
-      }
-      continue;
-    }
-    // `function X() { ... }` — plain function declaration, not memoised.
-    if (isNodeOfType(root, "FunctionDeclaration") && root.id) {
-      registry.set(root.id.name, "not-memoised");
-      continue;
-    }
-  }
-  return registry;
-};
-
 // Port of `oxc_linter::rules::react_perf::jsx_no_new_function_as_prop`.
 // Inline-expression coverage only — see jsx-no-new-array-as-prop's
 // LIMITATION note for the scope-analysis cases (`const x = () => {};
@@ -842,10 +768,6 @@ export const jsxNoNewFunctionAsProp = defineRule<Rule>({
   create: (context) => {
     const isTestlikeFile = isTestlikeFilename(context.getFilename?.());
     let memoRegistry: Map<string, MemoStatus> | null = null;
-    const lookupMemoStatus = (componentName: string): MemoStatus => {
-      if (!memoRegistry) return "unknown";
-      return memoRegistry.get(componentName) ?? "unknown";
-    };
     return {
       Program(node: EsTreeNodeOfType<"Program">) {
         memoRegistry = buildSameFileMemoRegistry(node as EsTreeNode);
@@ -866,14 +788,10 @@ export const jsxNoNewFunctionAsProp = defineRule<Rule>({
         // change, new function references included. The wrapper
         // pattern is unactionable noise here.
         const parentJsxOpening = node.parent;
-        if (
-          parentJsxOpening &&
-          isNodeOfType(parentJsxOpening, "JSXOpeningElement") &&
-          isNodeOfType(parentJsxOpening.name as EsTreeNode, "JSXIdentifier")
-        ) {
-          const componentName = (parentJsxOpening.name as EsTreeNodeOfType<"JSXIdentifier">).name;
-          if (lookupMemoStatus(componentName) === "not-memoised") return;
-        }
+        const openingName = parentJsxOpening && isNodeOfType(parentJsxOpening, "JSXOpeningElement")
+          ? (parentJsxOpening.name as EsTreeNode)
+          : null;
+        if (memoStatusForJsxOpeningName(memoRegistry, openingName) === "not-memoised") return;
         // One-shot lifecycle handlers (onMount / onError / onClose /
         // etc.) and render-prop slots (`fallback`, `render*`, `*Render`,
         // `*Renderer`, etc.) accept inline functions by design — they
